requirements.txt
Flask==2.0.1
Flask-CacheBuster==1.0.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
gradio==2.2.8
gunicorn==20.1.0
h5py==3.1.0
# ipython==7.24.0
# ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
Jinja2==3.0.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.2.54
Pillow==8.2.0
PyQtWebEngine==5.12.1
requests==2.25.1
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
